Determining the best tariff for a telecom company
A preliminary analysis of the use of tariffs on a sample of customers was carried out, the behavior of customers when using the operator's services was analyzed, and optimal sets of services for users were recommended. The data were preprocessed and analyzed. The hypotheses about the difference in the revenue of subscribers of different tariffs and the difference in the revenue of subscribers from Moscow and other regions were tested.