Analysis of the computer games market
The parameters that determine the success of the game in different regions of the world are revealed. Based on this, a report was prepared for a computer games store to plan advertising campaigns. Data preprocessing and analysis were carried out. The current period for analysis has been selected. Compiled portraits of users in each region. Hypotheses tested: the average user ratings of the Xbox One and PC platforms are the same; the average user ratings for the Action and Sports genres are different. The analysis used Student's t-test for independent samples.