Analysis of the mobile food application
Hypotheses were prioritized by ICE and RICE frameworks. Then I analyzed the results of the A / B test, built graphs of cumulative revenue, average check, conversion by group, and then calculated the statistical significance of differences in conversions and average checks for raw and cleaned data. Based on the analysis, I decided that it was not advisable to continue the test.